third_party/WebKit/LayoutTests/vr/latest/getDevices_zero_devices.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../resources/mock-vr-service.js"></script>
<script src="../resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.vr.getDevices().then( (devices) => {
    t.step( () => {
      assert_true(devices != null);
      assert_equals(devices.length, 0);
    }, "getDevices returned correct results");
  }, (err) => {
    t.step( () => {
      assert_unreached("getDevices rejected");
    });
  }).then( () => {
    t.done();
  });
}, [],
"navigator.vr.getDevices properly returns zero devices");

</script>
